Refactor main loop (#296)
* merge Simulator into openMalaria.cpp

* remove while(true) in openMalaria.cpp

* remove phase enum, make phase transitions explicit

* make population and transmission local variables (instead of global)

* remove all global variables from openMalaria.cpp

* merge Parameters.cpp into Parameters and remove boost dependency
