+++ model/util/random.cpp
@@ -1,0 +1,7 @@
+#include "random.h"
+
+namespace OM{
+    namespace util {
+        MasterRng master_RNG(0, 0);
+    }
+}
